Update types and implementation of the KNNRegressor algorithm.

+++ src/neighbors/knn_regressor/mod.rs
@@ -39,9 +39,9 @@ use parameters::{
 
 define_and_impl! {
     feature_type: f32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixF32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Euclidian<f32>,
     parameters_type: KNNRegressorF32EuclidianF32Parameters,
     distance_export_type: EuclidianF32
@@ -49,9 +49,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: f32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixF32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Mahalanobis<f32, DenseMatrix<f64>>,
     parameters_type: KNNRegressorF32MahalanobisF32Parameters,
     distance_export_type: MahalanobisF32
@@ -59,9 +59,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: f32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixF32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Manhattan<f32>,
     parameters_type: KNNRegressorF32ManhattanF32Parameters,
     distance_export_type: ManhattanF32
@@ -69,9 +69,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: f32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixF32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Minkowski<f32>,
     parameters_type: KNNRegressorF32MinkowskiF32Parameters,
     distance_export_type: MinkowskiF32
@@ -79,9 +79,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: f64,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixF64,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Euclidian<f64>,
     parameters_type: KNNRegressorF64EuclidianF64Parameters,
     distance_export_type: EuclidianF64
@@ -89,9 +89,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: f64,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixF64,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Mahalanobis<f64, DenseMatrix<f64>>,
     parameters_type: KNNRegressorF64MahalanobisF64Parameters,
     distance_export_type: MahalanobisF64
@@ -99,9 +99,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: f64,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixF64,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Manhattan<f64>,
     parameters_type: KNNRegressorF64ManhattanF64Parameters,
     distance_export_type: ManhattanF64
@@ -109,9 +109,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: f64,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixF64,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Minkowski<f64>,
     parameters_type: KNNRegressorF64MinkowskiF64Parameters,
     distance_export_type: MinkowskiF64
@@ -119,9 +119,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: i32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixI32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Euclidian<i32>,
     parameters_type: KNNRegressorI32EuclidianI32Parameters,
     distance_export_type: EuclidianI32
@@ -129,9 +129,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: i32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixI32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Hamming<i32>,
     parameters_type: KNNRegressorI32HammingI32Parameters,
     distance_export_type: HammingI32
@@ -139,9 +139,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: i32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixI32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Manhattan<i32>,
     parameters_type: KNNRegressorI32ManhattanI32Parameters,
     distance_export_type: ManhattanI32
@@ -149,9 +149,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: i32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixI32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Minkowski<i32>,
     parameters_type: KNNRegressorI32MinkowskiI32Parameters,
     distance_export_type: MinkowskiI32
@@ -159,9 +159,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: i64,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixI64,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Euclidian<i64>,
     parameters_type: KNNRegressorI64EuclidianI64Parameters,
     distance_export_type: EuclidianI64
@@ -169,9 +169,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: i64,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixI64,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Manhattan<i64>,
     parameters_type: KNNRegressorI64ManhattanI64Parameters,
     distance_export_type: ManhattanI64
@@ -179,9 +179,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: i64,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixI64,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Minkowski<i64>,
     parameters_type: KNNRegressorI64MinkowskiI64Parameters,
     distance_export_type: MinkowskiI64
@@ -189,9 +189,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: u16,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixU16,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Euclidian<u16>,
     parameters_type: KNNRegressorU16EuclidianU16Parameters,
     distance_export_type: EuclidianU16
@@ -199,9 +199,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: u16,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixU16,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Hamming<u16>,
     parameters_type: KNNRegressorU16HammingU16Parameters,
     distance_export_type: HammingU16
@@ -209,9 +209,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: u32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixU32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Euclidian<u32>,
     parameters_type: KNNRegressorU32EuclidianU32Parameters,
     distance_export_type: EuclidianU32
@@ -219,9 +219,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: u32,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixU32,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Manhattan<u32>,
     parameters_type: KNNRegressorU32ManhattanU32Parameters,
     distance_export_type: ManhattanU32
@@ -229,9 +229,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: u64,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixU64,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Euclidian<u64>,
     parameters_type: KNNRegressorU64EuclidianU64Parameters,
     distance_export_type: EuclidianU64
@@ -239,9 +239,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: u64,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixU64,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Manhattan<u64>,
     parameters_type: KNNRegressorU64ManhattanU64Parameters,
     distance_export_type: ManhattanU64
@@ -249,9 +249,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: u8,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixU8,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Euclidian<u8>,
     parameters_type: KNNRegressorU8EuclidianU8Parameters,
     distance_export_type: EuclidianU8
@@ -259,9 +259,9 @@ define_and_impl! {
 
 define_and_impl! {
     feature_type: u8,
-    target_type: i32,
+    target_type: f32,
     matrix_type: DenseMatrixU8,
-    array_type: Int32Array,
+    array_type: Float32Array,
     distance_type: Hamming<u8>,
     parameters_type: KNNRegressorU8HammingU8Parameters,
     distance_export_type: HammingU8
